+++ hint_accountant/src/main.rs
@@ -51,12 +51,11 @@ fn run() {
     }
     let mut vm = VirtualMachine::new(false, false);
     let mut hint_executor = BuiltinHintProcessor::new_empty();
-    let (ap_tracking_data, reference_ids, references, mut exec_scopes, constants) = (
+    let (ap_tracking_data, reference_ids, references, mut exec_scopes) = (
         ApTracking::default(),
         HashMap::new(),
         Vec::new(),
         ExecutionScopes::new(),
-        HashMap::new(),
     );
     let missing_hints: HashSet<_> = whitelists
         .into_iter()
@@ -64,10 +63,16 @@ fn run() {
         .map(|ahe| ahe.hint_lines.join("\n"))
         .filter(|h| {
             let hint_data = hint_executor
-                .compile_hint(h, &ap_tracking_data, &reference_ids, &references)
+                .compile_hint(
+                    h,
+                    &ap_tracking_data,
+                    &reference_ids,
+                    &references,
+                    Default::default(),
+                )
                 .expect("this implementation is infallible");
             matches!(
-                hint_executor.execute_hint(&mut vm, &mut exec_scopes, &hint_data, &constants,),
+                hint_executor.execute_hint(&mut vm, &mut exec_scopes, &hint_data),
                 Err(HintError::UnknownHint(_)),
             )
         })